+++ controllers/articles.controllers.js
@@ -60,16 +60,7 @@ function patchArticleVotes(request, response, next) {
     });
 }
 
-function deleteCommentByID(request, response, next) {
-  const commentID = request.params.comment_id;
-  return deleteComment(commentID)
-    .then(() => {
-      response.status(204).send();
-    })
-    .catch((err) => {
-      return next(err);
-    });
-}
+
 
 module.exports = {
   getArticlesId,
@@ -77,5 +68,4 @@ module.exports = {
   getCommentsFromArticle,
   postCommentToArticle,
   patchArticleVotes,
-  deleteCommentByID
 };
