+++ models/articles.models.js
@@ -92,15 +92,7 @@ function updateArticleVotes(articleID, requestBody) {
     });
 }
 
-function deleteComment(commentID) {
-  return db
-    .query(
-      `DELETE FROM comments
-    WHERE comment_id = $1
-    RETURNING *;`,
-      [commentID]
-    )
-}
+
 
 module.exports = {
   selectArticlesId,
@@ -108,5 +100,4 @@ module.exports = {
   selectCommentsFromArticle,
   insertCommentToArticle,
   updateArticleVotes,
-  deleteComment,
 };
